Added support for multiple javascript rule sets to wasm_bindgen package
