+++ .github/workflows/formatting.yaml
@@ -10,7 +10,7 @@ on:
       - synchronize
 
 jobs:
-  clang-format-checking:
+  code-format-checks:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
@@ -19,3 +19,13 @@ jobs:
           source: '.'
           extensions: 'h,c,cc,proto'
           clangFormatVersion: 14
+      - uses: actionsx/prettier@v2
+        with:
+          args: --config "${{ github.workspace }}/.prettierrc.toml" --write "**/*.{js,ts}"
+      # Prettier has no diff view so we must make one ourselves
+      # https://github.com/prettier/prettier/issues/6885
+      - run: |
+          git diff
+          if [[ -n "$(git status --porcelain)" ]]; then
+            exit 1
+          fi
