+++ docs/rust_wasm_bindgen.vm
@@ -1,18 +1,38 @@
 #[[
 ## Overview
 
-To build a `rust_binary` for `wasm32-unknown-unknown` target add the `--platforms=@rules_rust//rust/platform:wasm` flag.
+Bazel rules for generating wasm modules for Javascript using [wasm-bindgen][wb].
 
-```command
-bazel build @examples//hello_world_wasm --platforms=@rules_rust//rust/platform:wasm
+## Setup
+
+To begin using the `wasm-bindgen` rules, users can load the necessary dependencies
+in their workspace by adding the following to their `WORKSPACE.bazel` file.
+
+```starlark
+load("@rules_rust//wasm_bindgen:repositories.bzl", "rust_wasm_bindgen_dependencies", "rust_wasm_bindgen_register_toolchains")
+
+rust_wasm_bindgen_dependencies()
+
+rust_wasm_bindgen_register_toolchains()
 ```
 
-To build a `rust_binary` for `wasm32-wasi` target add the `--platforms=@rules_rust//rust/platform:wasi` flag.
+This should enable users to start using the [rust_wasm_bindgen](#rust_wasm_bindgen)
+rule. However, it's common to want to control the version of `wasm-bindgen` in the
+workspace instead of relying on the one provided by `rules_rust`. In this case, users
+should avoid calling `rust_wasm_bindgen_register_toolchains` and instead use the
+[rust_wasm_bindgen_toolchain](#rust_wasm_bindgen_toolchain) rule to define their own
+toolchains to register in the workspace.
 
-```command
-bazel build @examples//hello_world_wasm --platforms=@rules_rust//rust/platform:wasi
-```
+### Interfacing with Javascript rules
 
-`rust_wasm_bindgen` will automatically transition to the `wasm` platform and can be used when
-building WebAssembly code for the host target.
+While it's recommended for users to mantain their own , in the
+`@rules_rust//wasm_bindgen` package there exists interface sub-packages for various
+Javascript Bazel rules. E.g. `build_bazel_rules_nodejs` or `aspect_rules_js`. The
+rules defined there are a more convenient way to use `rust_wasm_bindgen` with the
+associated javascript rules due to the inclusion of additional providers. Each
+directory contains a `defs.bzl` file that defines the different variants of
+`rust_wasm_bindgen`. (e.g. `nodejs_rust_wasm_bindgen` for the `rules_nodejs` submodule).
+
+
+[wb]: https://github.com/rustwasm/wasm-bindgen
 ]]#
