+++ examples/wasm_bindgen/main.rs
