+++ wasm_bindgen/BUILD.bazel
@@ -1,17 +1,10 @@
 load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
-load("//wasm_bindgen:wasm_bindgen.bzl", "rust_wasm_bindgen_toolchain")
+load("//wasm_bindgen:defs.bzl", "rust_wasm_bindgen_toolchain")
 
 package(default_visibility = ["//visibility:public"])
 
 toolchain_type(
     name = "toolchain_type",
-)
-
-alias(
-    name = "wasm_bindgen_toolchain",
-    actual = "toolchain_type",
-    deprecation = "instead use `@rules_rust//wasm_bindgen:toolchain_type`",
-    tags = ["manual"],
 )
 
 bzl_library(
