+++ wasm_bindgen/providers.bzl
@@ -1,10 +1,10 @@
-"""A module for re-exporting the providers used by the rust_wasm_bindgen rule"""
+"""Rust WASM bindgen providers"""
 
-load(
-    "@rules_nodejs//nodejs:providers.bzl",
-    _DeclarationInfo = "DeclarationInfo",
-    _JSModuleInfo = "JSModuleInfo",
+RustWasmBindgenInfo = provider(
+    doc = "Info about wasm-bindgen outputs.",
+    fields = {
+        "js": "Depset[File]: The Javascript files produced by `wasm-bindgen`.",
+        "ts": "Depset[File]: The Typescript files produced by `wasm-bindgen`.",
+        "wasm": "File: The `.wasm` file generated by `wasm-bindgen`.",
+    },
 )
-
-DeclarationInfo = _DeclarationInfo
-JSModuleInfo = _JSModuleInfo
