+++ wasm_bindgen/rules_js/BUILD.bazel
@@ -1,0 +1,1 @@
+"""Rust WASM-bindgen rules for interfacing with aspects/rules_js"""
